style(tailwind.config.ts): add new style for '[class~="lead"]' to set font size and line height
The changes were made to the tailwind.config.ts file to add a new style for the '[class~="lead"]' selector. This style sets the font size using the 'clamp' function and sets the line height to 1.4. This change was made to improve the styling of the 'lead' class in the application.